requirements.txt
requests
pymongo
python-dotenv
pillow